add the retry to upload/download bucket operations
Add backoff retry for a single object storage request, except Range and Iter.
Error handler splits errors on net/http and others, and replies the request to the object storage for the former.
