+++ pkg/objstore/objstore.go
@@ -1,18 +1,20 @@
 package objstore
 
 import (
+	"bytes"
 	"context"
 	"io"
+	"net"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
+	"github.com/cenkalti/backoff"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/improbable-eng/thanos/pkg/runutil"
 	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus"
 )
 
 // Bucket provides read and write access to an object storage bucket.
@@ -83,7 +85,19 @@ func UploadFile(ctx context.Context, logger log.Logger, bkt Bucket, src, dst str
 	}
 	defer runutil.CloseWithLogOnErr(logger, r, "close file %s", src)
 
-	if err := bkt.Upload(ctx, dst, r); err != nil {
+	errorNotifier := func(err error, d time.Duration) {
+		level.Warn(logger).Log("msg", "unsuccessful attempt uploading", "src", src, "dst", dst, "err", err, "next", d.String())
+	}
+
+	do := func() error {
+		if err := bkt.Upload(ctx, dst, r); err != nil {
+			return handleErrors(errors.Wrapf(err, "upload file %s", dst))
+		}
+		return nil
+	}
+
+	// TODO: should we clean an uploaded file before retry?
+	if err := backoff.RetryNotify(do, getCustomBackoff(ctx), errorNotifier); err != nil {
 		return errors.Wrapf(err, "upload file %s as %s", src, dst)
 	}
 	return nil
@@ -93,14 +107,59 @@ func UploadFile(ctx context.Context, logger log.Logger, bkt Bucket, src, dst str
 const DirDelim = "/"
 
 // DeleteDir removes all objects prefixed with dir from the bucket.
-func DeleteDir(ctx context.Context, bkt Bucket, dir string) error {
-	return bkt.Iter(ctx, dir, func(name string) error {
+func DeleteDir(ctx context.Context, bucket Bucket, dir string) error {
+	// Delete directory.
+	do := func() error {
+		if err := deleteDir(ctx, bucket, dir); err != nil {
+			return handleErrors(errors.Wrapf(err, "delete block dir %s", dir))
+		}
+		return nil
+	}
+
+	if err := backoff.Retry(do, getCustomBackoff(ctx)); err != nil {
+		return errors.Wrapf(err, "delete block dir %s", dir)
+	}
+	return nil
+}
+
+// deleteDir is an internal function to delete directory recursively without retries.
+func deleteDir(ctx context.Context, bucket Bucket, dir string) error {
+	return bucket.Iter(ctx, dir, func(name string) error {
 		// If we hit a directory, call DeleteDir recursively.
 		if strings.HasSuffix(name, DirDelim) {
-			return DeleteDir(ctx, bkt, name)
+			return deleteDir(ctx, bucket, name)
 		}
-		return bkt.Delete(ctx, name)
+		return bucket.Delete(ctx, name)
 	})
+}
+
+// GetFile downloads file from bucket and returns bytes.
+func GetFile(ctx context.Context, logger log.Logger, bucket BucketReader, src string) ([]byte, error) {
+	buf := bytes.NewBuffer(nil)
+
+	errorNotifier := func(err error, d time.Duration) {
+		level.Warn(logger).Log("msg", "unsuccessful attempt to download", "err", err, "next", d.String())
+	}
+
+	do := func() error {
+		rc, err := bucket.Get(ctx, src)
+		if err != nil {
+			return handleErrors(errors.Wrapf(err, "download file %s from bucket", src))
+		}
+		defer runutil.CloseWithLogOnErr(logger, rc, "download file")
+
+		if _, err := io.Copy(buf, rc); err != nil {
+			return handleErrors(errors.Wrap(err, "download file"))
+		}
+
+		return nil
+	}
+
+	if err := backoff.RetryNotify(do, getCustomBackoff(ctx), errorNotifier); err != nil {
+		return nil, errors.Wrapf(err, "download file %s", src)
+	}
+
+	return buf.Bytes(), nil
 }
 
 // DownloadFile downloads the src file from the bucket to dst. If dst is an existing
@@ -115,12 +174,6 @@ func DownloadFile(ctx context.Context, logger log.Logger, bkt BucketReader, src,
 		return err
 	}
 
-	rc, err := bkt.Get(ctx, src)
-	if err != nil {
-		return errors.Wrap(err, "get file")
-	}
-	defer runutil.CloseWithLogOnErr(logger, rc, "download block's file reader")
-
 	f, err := os.Create(dst)
 	if err != nil {
 		return errors.Wrap(err, "create file")
@@ -134,232 +187,104 @@ func DownloadFile(ctx context.Context, logger log.Logger, bkt BucketReader, src,
 			}
 		}
 	}()
-	if _, err = io.Copy(f, rc); err != nil {
-		return errors.Wrap(err, "copy object to file")
+
+	errorNotifier := func(err error, d time.Duration) {
+		level.Warn(logger).Log("msg", "unsuccessful attempt to download", "err", err, "next", d.String())
+	}
+
+	do := func() error {
+		rc, err := bkt.Get(ctx, src)
+		if err != nil {
+			return handleErrors(errors.Wrap(err, "get file"))
+		}
+		defer runutil.CloseWithLogOnErr(logger, rc, "download block's file reader")
+
+		if _, err = io.Copy(f, rc); err != nil {
+			return handleErrors(errors.Wrap(err, "copy object to file"))
+		}
+		return nil
+	}
+	err = backoff.RetryNotify(do, getCustomBackoff(ctx), errorNotifier)
+	if err != nil {
+		return errors.Wrapf(err, "failed to download file %s", src)
 	}
 	return nil
 }
 
 // DownloadDir downloads all object found in the directory into the local directory.
-func DownloadDir(ctx context.Context, logger log.Logger, bkt BucketReader, src, dst string) error {
-	if err := os.MkdirAll(dst, 0777); err != nil {
+func DownloadDir(ctx context.Context, logger log.Logger, bucket BucketReader, src, dst string) error {
+	var (
+		err             error
+		downloadedFiles []string
+	)
+	err = os.MkdirAll(dst, 0777)
+	if err != nil {
 		return errors.Wrap(err, "create dir")
 	}
 
-	var downloadedFiles []string
-	if err := bkt.Iter(ctx, src, func(name string) error {
-		if strings.HasSuffix(name, DirDelim) {
-			return DownloadDir(ctx, logger, bkt, name, filepath.Join(dst, filepath.Base(name)))
+	defer func() {
+		if err != nil {
+			// Best-effort cleanup if the download failed.
+			for _, f := range downloadedFiles {
+				if rErr := os.Remove(f); rErr != nil {
+					level.Warn(logger).Log("msg", "failed to remove file on partial dir download error", "file", f, "err", rErr)
+				}
+			}
 		}
-		if err := DownloadFile(ctx, logger, bkt, name, dst); err != nil {
+	}()
+
+	// TODO(xjewer): handle iterations, what's the best effort?
+	err = bucket.Iter(ctx, src, func(name string) error {
+		if strings.HasSuffix(name, DirDelim) {
+			return DownloadDir(ctx, logger, bucket, name, filepath.Join(dst, filepath.Base(name)))
+		}
+		if err := DownloadFile(ctx, logger, bucket, name, dst); err != nil {
 			return err
 		}
 
 		downloadedFiles = append(downloadedFiles, dst)
 		return nil
-	}); err != nil {
-		// Best-effort cleanup if the download failed.
-		for _, f := range downloadedFiles {
-			if rerr := os.Remove(f); rerr != nil {
-				level.Warn(logger).Log("msg", "failed to remove file on partial dir download error", "file", f, "err", rerr)
-			}
-		}
-		return err
-	}
+	})
 
 	return nil
 }
 
-// BucketWithMetrics takes a bucket and registers metrics with the given registry for
-// operations run against the bucket.
-func BucketWithMetrics(name string, b Bucket, r prometheus.Registerer) Bucket {
-	bkt := &metricBucket{
-		bkt: b,
+// Exists checks if file exists with backoff retries.
+func Exists(ctx context.Context, logger log.Logger, bucket BucketReader, src string) (bool, error) {
+	var ok bool
 
-		ops: prometheus.NewCounterVec(prometheus.CounterOpts{
-			Name:        "thanos_objstore_bucket_operations_total",
-			Help:        "Total number of operations against a bucket.",
-			ConstLabels: prometheus.Labels{"bucket": name},
-		}, []string{"operation"}),
-
-		opsFailures: prometheus.NewCounterVec(prometheus.CounterOpts{
-			Name:        "thanos_objstore_bucket_operation_failures_total",
-			Help:        "Total number of operations against a bucket that failed.",
-			ConstLabels: prometheus.Labels{"bucket": name},
-		}, []string{"operation"}),
-
-		opsDuration: prometheus.NewHistogramVec(prometheus.HistogramOpts{
-			Name:        "thanos_objstore_bucket_operation_duration_seconds",
-			Help:        "Duration of operations against the bucket",
-			ConstLabels: prometheus.Labels{"bucket": name},
-			Buckets:     []float64{0.005, 0.01, 0.02, 0.04, 0.08, 0.15, 0.3, 0.6, 1, 1.5, 2.5, 5, 10, 20, 30},
-		}, []string{"operation"}),
-		lastSuccessfullUploadTime: prometheus.NewGauge(prometheus.GaugeOpts{
-			Name:        "thanos_objstore_bucket_last_successful_upload_time",
-			Help:        "Second timestamp of the last successful upload to the bucket.",
-			ConstLabels: prometheus.Labels{"bucket": name}}),
+	errorNotifier := func(err error, d time.Duration) {
+		level.Warn(logger).Log("msg", "unsuccessful attempt getting the file/dir", "src", src, "err", err, "next", d.String())
 	}
-	if r != nil {
-		r.MustRegister(bkt.ops, bkt.opsFailures, bkt.opsDuration, bkt.lastSuccessfullUploadTime)
+
+	do := func() error {
+		var err error
+		ok, err = bucket.Exists(ctx, src)
+		if err != nil {
+			return handleErrors(errors.Wrap(err, "check exists"))
+		}
+		return nil
 	}
-	return bkt
-}
 
-type metricBucket struct {
-	bkt Bucket
-
-	ops                       *prometheus.CounterVec
-	opsFailures               *prometheus.CounterVec
-	opsDuration               *prometheus.HistogramVec
-	lastSuccessfullUploadTime prometheus.Gauge
-}
-
-func (b *metricBucket) Iter(ctx context.Context, dir string, f func(name string) error) error {
-	const op = "iter"
-
-	err := b.bkt.Iter(ctx, dir, f)
+	err := backoff.RetryNotify(do, getCustomBackoff(ctx), errorNotifier)
 	if err != nil {
-		b.opsFailures.WithLabelValues(op).Inc()
+		return ok, errors.Wrapf(err, "failed to check file %s", src)
 	}
-	b.ops.WithLabelValues(op).Inc()
 
-	return err
+	return ok, nil
 }
 
-func (b *metricBucket) Get(ctx context.Context, name string) (io.ReadCloser, error) {
-	const op = "get"
-	b.ops.WithLabelValues(op).Inc()
-
-	rc, err := b.bkt.Get(ctx, name)
-	if err != nil {
-		b.opsFailures.WithLabelValues(op).Inc()
-		return nil, err
-	}
-	rc = newTimingReadCloser(
-		rc,
-		op,
-		b.opsDuration,
-		b.opsFailures,
-	)
-
-	return rc, nil
-}
-
-func (b *metricBucket) GetRange(ctx context.Context, name string, off, length int64) (io.ReadCloser, error) {
-	const op = "get_range"
-	b.ops.WithLabelValues(op).Inc()
-
-	rc, err := b.bkt.GetRange(ctx, name, off, length)
-	if err != nil {
-		b.opsFailures.WithLabelValues(op).Inc()
-		return nil, err
-	}
-	rc = newTimingReadCloser(
-		rc,
-		op,
-		b.opsDuration,
-		b.opsFailures,
-	)
-
-	return rc, nil
-}
-
-func (b *metricBucket) Exists(ctx context.Context, name string) (bool, error) {
-	const op = "exists"
-	start := time.Now()
-
-	ok, err := b.bkt.Exists(ctx, name)
-	if err != nil {
-		b.opsFailures.WithLabelValues(op).Inc()
-	}
-	b.ops.WithLabelValues(op).Inc()
-	b.opsDuration.WithLabelValues(op).Observe(time.Since(start).Seconds())
-
-	return ok, err
-}
-
-func (b *metricBucket) Upload(ctx context.Context, name string, r io.Reader) error {
-	const op = "upload"
-	start := time.Now()
-
-	err := b.bkt.Upload(ctx, name, r)
-	if err != nil {
-		b.opsFailures.WithLabelValues(op).Inc()
-	} else {
-		//TODO: Use SetToCurrentTime() once we update the Prometheus client_golang
-		b.lastSuccessfullUploadTime.Set(float64(time.Now().UnixNano()) / 1e9)
-	}
-	b.ops.WithLabelValues(op).Inc()
-	b.opsDuration.WithLabelValues(op).Observe(time.Since(start).Seconds())
-
-	return err
-}
-
-func (b *metricBucket) Delete(ctx context.Context, name string) error {
-	const op = "delete"
-	start := time.Now()
-
-	err := b.bkt.Delete(ctx, name)
-	if err != nil {
-		b.opsFailures.WithLabelValues(op).Inc()
-	}
-	b.ops.WithLabelValues(op).Inc()
-	b.opsDuration.WithLabelValues(op).Observe(time.Since(start).Seconds())
-
-	return err
-}
-
-func (b *metricBucket) IsObjNotFoundErr(err error) bool {
-	return b.bkt.IsObjNotFoundErr(err)
-}
-
-func (b *metricBucket) Close() error {
-	return b.bkt.Close()
-}
-
-func (b *metricBucket) Name() string {
-	return b.bkt.Name()
-}
-
-type timingReadCloser struct {
-	io.ReadCloser
-
-	ok       bool
-	start    time.Time
-	op       string
-	duration *prometheus.HistogramVec
-	failed   *prometheus.CounterVec
-}
-
-func newTimingReadCloser(rc io.ReadCloser, op string, dur *prometheus.HistogramVec, failed *prometheus.CounterVec) *timingReadCloser {
-	// Initialize the metrics with 0.
-	dur.WithLabelValues(op)
-	failed.WithLabelValues(op)
-	return &timingReadCloser{
-		ReadCloser: rc,
-		ok:         true,
-		start:      time.Now(),
-		op:         op,
-		duration:   dur,
-		failed:     failed,
-	}
-}
-
-func (rc *timingReadCloser) Close() error {
-	err := rc.ReadCloser.Close()
-	rc.duration.WithLabelValues(rc.op).Observe(time.Since(rc.start).Seconds())
-	if rc.ok && err != nil {
-		rc.failed.WithLabelValues(rc.op).Inc()
-		rc.ok = false
+// handleErrors handles net error and wraps in permanent otherwise.
+func handleErrors(err error) error {
+	cErr := errors.Cause(err)
+	if _, ok := cErr.(net.Error); !ok {
+		return backoff.Permanent(errors.Wrap(err, "permanent error"))
 	}
 	return err
 }
 
-func (rc *timingReadCloser) Read(b []byte) (n int, err error) {
-	n, err = rc.ReadCloser.Read(b)
-	if rc.ok && err != nil && err != io.EOF {
-		rc.failed.WithLabelValues(rc.op).Inc()
-		rc.ok = false
-	}
-	return n, err
+// getCustomBackoff returns custom backoff to be used to retry operations.
+func getCustomBackoff(ctx context.Context) backoff.BackOff {
+	// TODO(xjewer): customize backoff with the data from context
+	return backoff.WithContext(backoff.NewExponentialBackOff(), ctx)
 }
